+++ src/Data/ToolDataLoader.cpp
@@ -40,7 +40,7 @@ bool ToolDataLoader::loadData(std::string toolDataPath)
 
         projectileNameToTypeMap[projectileData.name] = projectileIdx;
 
-        int sellValue = 0;
+        float sellValue = 0;
         if (jsonProjectileData.contains("sell-value")) sellValue = jsonProjectileData.at("sell-value");
 
         // Link with item data
@@ -118,7 +118,7 @@ bool ToolDataLoader::loadData(std::string toolDataPath)
             }
         }
 
-        int sellValue = 0;
+        float sellValue = 0;
         if (jsonToolData.contains("sell-value")) sellValue = jsonToolData.at("sell-value");
 
         loaded_toolData.push_back(toolData);
